requirements.txt
langchain
python-dotenv
langchain-community
langchain-core
langchain-groq
streamlit
sentence-transformers
pypdf
faiss-cpu
langchain-chroma
arxiv
wikipedia
duckduckgo-search
mysql-connector-python
SQLAlchemy